# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Spin API service
  description: HTTP service for controlling the spin service
  version: ""
host: localhost:8080
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /control/shutdown/{id}:
    post:
      tags:
      - spin-apiserver
      summary: control_shutdown spin-apiserver
      operationId: spin-apiserver#control_shutdown
      parameters:
      - name: id
        in: path
        description: ID of VM to shutdown
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
      schemes:
      - http
  /control/start/{id}:
    post:
      tags:
      - spin-apiserver
      summary: control_start spin-apiserver
      operationId: spin-apiserver#control_start
      parameters:
      - name: id
        in: path
        description: ID of VM to start
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
      schemes:
      - http
  /control/stop/{id}:
    post:
      tags:
      - spin-apiserver
      summary: control_stop spin-apiserver
      operationId: spin-apiserver#control_stop
      parameters:
      - name: id
        in: path
        description: ID of VM to stop
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
      schemes:
      - http
...
  /vm/create:
    post:
      tags:
      - spin-registry
      summary: vm_create spin-registry
      description: Create a VM
      operationId: spin-registry#vm_create
      parameters:
      - name: vm_create_request_body
        in: body
        required: true
        schema:
          $ref: '#/definitions/SpinRegistryVMCreateRequestBody'
          required:
          - name
          - cpus
          - memory
          - images
      responses:
        "200":
          description: OK response.
          schema:
            type: integer
            format: int64
      schemes:
      - http
  /vm/delete/{id}:
    post:
      tags:
      - spin-registry
      summary: vm_delete spin-registry
      description: Delete a VM by ID
      operationId: spin-registry#vm_delete
      parameters:
      - name: id
        in: path
        description: ID of VM to remove
        required: true
        type: integer
      responses:
        "200":
          description: OK response.
      schemes:
      - http
...
definitions:
  ImageRequestBody:
    title: ImageRequestBody
    type: object
    properties:
      cdrom:
        type: boolean
        description: Is this a cdrom image?
        example: false
      path:
        type: string
        description: Image path
        example: Sed necessitatibus et asperiores.
    example:
      cdrom: false
      path: Sequi officia animi fugiat.
    required:
    - path
    - cdrom
...
  SpinRegistryVMCreateRequestBody:
    title: SpinRegistryVMCreateRequestBody
    type: object
    properties:
      cpus:
        type: integer
        description: CPU count
        example: 743273099863973551
        format: int64
      images:
        type: array
        items:
          $ref: '#/definitions/ImageRequestBody'
        description: Image references
        example:
        - cdrom: false
          path: Corrupti et voluptatibus et et occaecati.
        - cdrom: false
          path: Corrupti et voluptatibus et et occaecati.
      memory:
        type: integer
        description: Memory (in megabytes)
        example: 12485697019927615723
        format: int64
      name:
        type: string
        description: Name of VM; does not need to be unique
        example: Aut tempore aliquam.
    example:
      cpus: 848682044187229832
      images:
      - cdrom: false
        path: Corrupti et voluptatibus et et occaecati.
      - cdrom: false
        path: Corrupti et voluptatibus et et occaecati.
      - cdrom: false
        path: Corrupti et voluptatibus et et occaecati.
      - cdrom: false
        path: Corrupti et voluptatibus et et occaecati.
      memory: 865562045241744601
      name: Natus neque qui iusto ex et.
    required:
    - name
    - cpus
    - memory
    - images
...